requirements.txt
dash>=2.14.0
dash-bootstrap-components>=1.5.0
plotly>=5.18.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
wordcloud>=1.9.0
matplotlib>=3.7.0
requests>=2.31.0
dash-iconify>=0.1.0
google-genai>=1.0.0
